test: enable MainViewFactory-integ-test
